+++ modules/project/src/main/java/com/falsepattern/zigbrains/project/execution/run/ZigExecConfigRun.java
@@ -37,7 +37,8 @@ public class ZigExecConfigRun extends ZigExecConfigBase<ZigExecConfigRun> {
     private ZigConfigEditor.FilePathConfigurable filePath = new ZigConfigEditor.FilePathConfigurable("filePath", "File Path");
     private ZigConfigEditor.CheckboxConfigurable colored = ZigConfigEditor.coloredConfigurable("colored");
     private ZigConfigEditor.OptimizationConfigurable optimization = new ZigConfigEditor.OptimizationConfigurable("optimization");
-    private ZigConfigEditor.ArgsConfigurable exeArgs = new ZigConfigEditor.ArgsConfigurable("exeArgs", "Arguments for the compile exe");
+    private ZigConfigEditor.ArgsConfigurable compilerArgs = new ZigConfigEditor.ArgsConfigurable("compilerArgs", "Extra compiler command line arguments");
+    private ZigConfigEditor.ArgsConfigurable exeArgs = new ZigConfigEditor.ArgsConfigurable("exeArgs", "Output program command line arguments");
     public ZigExecConfigRun(@NotNull Project project, @NotNull ConfigurationFactory factory) {
         super(project, factory, "Zig Run");
     }
@@ -51,6 +52,7 @@ public class ZigExecConfigRun extends ZigExecConfigBase<ZigExecConfigRun> {
         if (!debug || optimization.forced) {
             result.addAll(List.of("-O", optimization.level.name()));
         }
+        result.addAll(List.of(compilerArgs.args));
         if (!debug) {
             result.add("--");
             result.addAll(List.of(exeArgs.args));
@@ -68,6 +70,7 @@ public class ZigExecConfigRun extends ZigExecConfigBase<ZigExecConfigRun> {
         val clone = super.clone();
         clone.filePath = filePath.clone();
         clone.colored = colored.clone();
+        clone.compilerArgs = compilerArgs.clone();
         clone.optimization = optimization.clone();
         clone.exeArgs = exeArgs.clone();
         return clone;
@@ -75,7 +78,7 @@ public class ZigExecConfigRun extends ZigExecConfigBase<ZigExecConfigRun> {
 
     @Override
     public @NotNull List<ZigConfigEditor.ZigConfigurable<?>> getConfigurables() {
-        return CollectionUtil.concat(super.getConfigurables(), filePath, optimization, colored, exeArgs);
+        return CollectionUtil.concat(super.getConfigurables(), filePath, optimization, colored, compilerArgs, exeArgs);
     }
 
     @Override